LPS-57310 Move portletkeys to api bundle
