+++ modules/apps/dynamic-data-lists/dynamic-data-lists-service/src/com/liferay/dynamic/data/lists/upgrade/v1_0_0/UpgradeLastPublishDate.java
@@ -14,6 +14,8 @@
 
 package com.liferay.dynamic.data.lists.upgrade.v1_0_0;
 
+import com.liferay.dynamic.data.lists.constants.DDLPortletKeys;
+
 /**
  * @author Levente Hudak
  */
