+++ modules/apps/dynamic-data-lists/dynamic-data-lists-test/build.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE project>
 
 <project>
-	<property name="import.shared" value="dynamic-data-lists-web" />
+	<property name="import.shared" value="dynamic-data-lists-api,dynamic-data-lists-web" />
 
 	<import file="../../../../tools/sdk/build-common-osgi-plugin.xml" />
 </project>
