+++ modules/apps/dynamic-data-lists/dynamic-data-lists-web/src/META-INF/resources/init.jsp
@@ -27,6 +27,7 @@ taglib uri="http://liferay.com/tld/ui" prefix="liferay-ui" %><%@
 taglib uri="http://liferay.com/tld/util" prefix="liferay-util" %>
 
 <%@ page import="com.liferay.dynamic.data.lists.constants.DDLActionKeys" %><%@
+page import="com.liferay.dynamic.data.lists.constants.DDLPortletKeys" %><%@
 page import="com.liferay.dynamic.data.lists.constants.DDLWebKeys" %><%@
 page import="com.liferay.dynamic.data.lists.exception.NoSuchRecordException" %><%@
 page import="com.liferay.dynamic.data.lists.exception.NoSuchRecordSetException" %><%@
@@ -43,7 +44,6 @@ page import="com.liferay.dynamic.data.lists.service.DDLRecordVersionServiceUtil"
 page import="com.liferay.dynamic.data.lists.service.permission.DDLPermission" %><%@
 page import="com.liferay.dynamic.data.lists.service.permission.DDLRecordSetPermission" %><%@
 page import="com.liferay.dynamic.data.lists.util.DDLUtil" %><%@
-page import="com.liferay.dynamic.data.lists.web.constants.DDLPortletKeys" %><%@
 page import="com.liferay.dynamic.data.lists.web.context.DDLDisplayContext" %><%@
 page import="com.liferay.dynamic.data.lists.web.search.RecordSetDisplayTerms" %><%@
 page import="com.liferay.dynamic.data.lists.web.search.RecordSetSearch" %><%@
